Traces list in side-panel
- Split file SummaryPanelNode => (SummaryPanelNode + SummaryPanelNodeMetrics)
- When traces are enabled & service is selected, show a tabbed view Traces/Metrics below contextual node info

Fixes https://github.com/kiali/kiali/issues/2840
